(VIDEO-672) Handle CallKit from AppDelegate
Instead of handling callkit from the ViewController we are now handling it from the App delegate in order to register an incoming call when receiving the push notification and the app is in background
